Consolidated intro-slogan, mission_text, and pet-text so they behave the same
Renamed all three into one element in CSS: slogan-text.
Replaces all CSS for text over an image so that the text over an image behaves the SAME regardless of the size of the page.
Allows for code-reuse and minimal instances of copy/paste.

Also renamed "tag_top" to "the_top" for all "go-to-top" buttons.